Put the list tool inside dynamic tool module
