+++ dynamic-simulation/dynamic-simulation-tool/src/test/java/com/powsybl/dynamicsimulation/tool/ListDynamicSimulationModelsToolTest.java
@@ -5,12 +5,8 @@
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  * SPDX-License-Identifier: MPL-2.0
  */
-package com.powsybl.dynamicsimulation.modelslist;
+package com.powsybl.dynamicsimulation.tool;
 
-import com.google.auto.service.AutoService;
-import com.powsybl.computation.ComputationManager;
-import com.powsybl.dynamicsimulation.*;
-import com.powsybl.iidm.network.Network;
 import com.powsybl.tools.Command;
 import com.powsybl.tools.Tool;
 import com.powsybl.tools.test.AbstractToolTest;
@@ -19,7 +15,6 @@ import org.junit.jupiter.api.Test;
 
 import java.io.IOException;
 import java.util.Collections;
-import java.util.concurrent.CompletableFuture;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNull;
@@ -28,25 +23,6 @@ import static org.junit.jupiter.api.Assertions.assertNull;
  * @author Laurent Issertial <laurent.issertial at rte-france.com>
  */
 class ListDynamicSimulationModelsToolTest extends AbstractToolTest {
-
-    @AutoService(DynamicSimulationProvider.class)
-    public static class DynamicSimulationProviderMock implements DynamicSimulationProvider {
-
-        @Override
-        public CompletableFuture<DynamicSimulationResult> run(Network network, DynamicModelsSupplier dynamicModelsSupplier, EventModelsSupplier eventModelsSupplier, CurvesSupplier curvesSupplier, String workingVariantId, ComputationManager computationManager, DynamicSimulationParameters parameters) {
-            return CompletableFuture.completedFuture(new DynamicSimulationResultImpl(true, null, Collections.emptyMap(), DynamicSimulationResult.emptyTimeLine()));
-        }
-
-        @Override
-        public String getName() {
-            return "Mock";
-        }
-
-        @Override
-        public String getVersion() {
-            return null;
-        }
-    }
 
     private final ListDynamicSimulationModelsTool tool = new ListDynamicSimulationModelsTool();
