+++ dynamic-simulation/pom.xml
@@ -25,6 +25,5 @@
         <module>dynamic-simulation-api</module>
         <module>dynamic-simulation-dsl</module>
         <module>dynamic-simulation-tool</module>
-        <module>dynamic-simulation-models-list</module>
     </modules>
 </project>
